requirements.txt
pyTelegramBotAPI
requests
python-dotenv
yandex-cloud